#31 Splitted s3 file tests into two separate files
